move ds packages one level up
