+++ graph/directed_test.go
