+++ list/node.go
