+++ list/stack.go
