+++ symboltable/helper.go
@@ -1,4 +1,4 @@
-package st
+package symboltable
 
 func max(a, b int) int {
 	if a < b {
