+++ symboltable/red_black.go
@@ -1,4 +1,4 @@
-package st
+package symboltable
 
 import (
 	"fmt"
